feat: add validation strategy REMOVE_DOTS_AND_REPLACE_WHITESPACE
This option is deprecated because specialized treatment of the IDs should ideally be performed upstream of the converter. This option is provided only to ensure backward compatibility with the old default behavior. Users are encouraged to perform ID manipulations before invoking the converter, preferably directly in the Network Graphic Editor (NGE).